+++ spec/spec_helper.rb
@@ -11,9 +11,9 @@ require "shoulda-matchers"
 
 require "active_record"
 require "pry"
+require "rspice"
 
 require_relative "../lib/flow/spec_helper"
-require "rspice"
 
 require "flow"