+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add `prefix` option to state accessor methods ([#203](https://github.com/Freshly/flow/pull/203), [#206](https://github.com/Freshly/flow/pull/206))
 - Use Set instead of Array for accessor trackers ([#205](https://github.com/Freshly/flow/pull/205))
 - Allow multiple state accessor definitions at once ([#207](https://github.com/Freshly/flow/pull/207))
+- Use Substance instead of duplicative base objects, remove Spicerack dependency ([#185](https://github.com/Freshly/flow/pull/185))
 
 ## 0.10.9